Move shared file up the tree
